move footer and header to separate folders
