+++ src/components/header/LogoV2.tsx
